requirements.txt
flask
flask-restful
mysql-connector
passlib
Flask-HTTPAuth
configparser
